+++ tests/test_matrix.cpp
@@ -5,7 +5,7 @@
 #include "test.h"
 
 TEST_CASE("Constructing and inspecting a 4x4 matrix") {
-   const std::vector<float> e = { 1,2,3,4,5.5,6.5,7.5,8.5,9,10,11,12,13.5,14.5,15.5,16.5 };
+   const std::vector<double> e = { 1,2,3,4,5.5,6.5,7.5,8.5,9,10,11,12,13.5,14.5,15.5,16.5 };
     Matrix m { e };
 
     REQUIRE(m.element(0, 0) == Catch::Approx(1.0));
@@ -18,7 +18,7 @@ TEST_CASE("Constructing and inspecting a 4x4 matrix") {
 }
 
 TEST_CASE("A 3x3 matrix ought to be representable") {
-    const std::vector<float> e1 = { -3, 5, 0, 1, -2, -7, 0, 1, 1};
+    const std::vector<double> e1 = { -3, 5, 0, 1, -2, -7, 0, 1, 1};
     const auto m = Matrix { e1 };
 
     REQUIRE(m.element(0, 0) == Catch::Approx(-3.0));
@@ -27,7 +27,7 @@ TEST_CASE("A 3x3 matrix ought to be representable") {
 }
 
 TEST_CASE("A 2x2 matrix ought to be representable") {
-    const std::vector<float> e = { -3, 5, 1, 2 };
+    const std::vector<double> e = { -3, 5, 1, 2 };
     const Matrix m { e };
     REQUIRE(m.element(0, 0) == Catch::Approx(-3.0));
     REQUIRE(m.element(0, 1) == Catch::Approx(5.0));
@@ -37,56 +37,56 @@ TEST_CASE("A 2x2 matrix ought to be representable") {
 
 
 TEST_CASE("Matrix equality with indentical matrices") {
-    const std::vector<float> e1 = { 1,2,3,4,5,6,7,8,9,8,7,6,5,4,3,2 };
+    const std::vector<double> e1 = { 1,2,3,4,5,6,7,8,9,8,7,6,5,4,3,2 };
     const Matrix a { e1};
-    const std::vector<float> e2 = { 1,2,3,4,5,6,7,8,9,8,7,6,5,4,3,2 };
+    const std::vector<double> e2 = { 1,2,3,4,5,6,7,8,9,8,7,6,5,4,3,2 };
     const Matrix b { e2};
     REQUIRE(a == b);
 }
 
 TEST_CASE("Matrix equality with different matrices") {
-    const std::vector<float> e1= { 1,2,3,4,5,6,7,8,9,8,7,6,5,4,3,2 };
+    const std::vector<double> e1= { 1,2,3,4,5,6,7,8,9,8,7,6,5,4,3,2 };
     const Matrix a { e1 };
-    const std::vector<float> e2 = { 2,3,4,5,6,7,8,9,8,7,6,5,4,3,2,1 };
+    const std::vector<double> e2 = { 2,3,4,5,6,7,8,9,8,7,6,5,4,3,2,1 };
     const Matrix b { e2 };
     REQUIRE(a != b);
 }
 
 TEST_CASE("Multiplying two matrices") {
-    const std::vector<float> e1 = { 1,2,3,4,5,6,7,8,9,8,7,6,5,4,3,2 };
+    const std::vector<double> e1 = { 1,2,3,4,5,6,7,8,9,8,7,6,5,4,3,2 };
     const Matrix a { e1};
-    const std::vector<float> e2 = { -2,1,2,3,3,2,1,-1,4,3,6,5,1,2,7,8 };
+    const std::vector<double> e2 = { -2,1,2,3,3,2,1,-1,4,3,6,5,1,2,7,8 };
     const Matrix b { e2};
 
     const Matrix result = a * b;
-    const std::vector<float> e3 = { 20,22,50,48,44,54,114,108,40,58,110,102,16,26,46,42 };
+    const std::vector<double> e3 = { 20,22,50,48,44,54,114,108,40,58,110,102,16,26,46,42 };
     const Matrix expected { e3 };
     REQUIRE_MATRICES(result, expected);
 }
 
 TEST_CASE("Multiplying a matrix by a tuple") {
-    const std::vector<float> e1 = { 1,2,3,4,2,4,4,2,8,6,4,1,0,0,0,1 };
+    const std::vector<double> e1 = { 1,2,3,4,2,4,4,2,8,6,4,1,0,0,0,1 };
     const auto a = Matrix { e1};
     const auto b = Tuple {1,2,3,1};
     REQUIRE_TUPLES(a * b, Tuple {18, 24, 33, 1});
 }
 
 TEST_CASE("Multiplying a matrix by the Identity Matrix") {
-    const std::vector<float> e1 = { 1,2,3,4,2,4,4,2,8,6,4,1,0,0,0,1 };
+    const std::vector<double> e1 = { 1,2,3,4,2,4,4,2,8,6,4,1,0,0,0,1 };
     const auto a = Matrix {e1};
     const auto b = Matrix::Identity();
     REQUIRE_MATRICES(a * b, a);
 }
 
 TEST_CASE("Multiplying a tuple by the Identity Matrix") {
-    float e1[16] = { 1,2,3,4,2,4,4,2,8,6,4,1,0,0,0,1 };
+    double e1[16] = { 1,2,3,4,2,4,4,2,8,6,4,1,0,0,0,1 };
     const auto a = Tuple { 1,2,3,4};
     const auto b = Matrix::Identity();
     REQUIRE_TUPLES(b * a, a);
 }
 
 TEST_CASE("Transposing a matrix A") {
-    const std::vector<float> e1 =
+    const std::vector<double> e1 =
     {
         0,9,3,0,
         9,8,0,8,
@@ -94,7 +94,7 @@ TEST_CASE("Transposing a matrix A") {
         0,0,5,8
     };
     const auto a = Matrix {e1};
-    const std::vector<float> e2 =
+    const std::vector<double> e2 =
     {
         0,9,1,0,
         9,8,8,0,
@@ -108,7 +108,7 @@ TEST_CASE("Transposing a matrix A") {
 
 TEST_CASE("Calculating the determinant of a 2x2 matrix")
 {
-    const std::vector<float> e =
+    const std::vector<double> e =
     {
         1,5,-3,2
     };
@@ -119,7 +119,7 @@ TEST_CASE("Calculating the determinant of a 2x2 matrix")
 
 TEST_CASE("A submatrix of a 3x3 is a 2x2")
 {
-    const std::vector<float> e =
+    const std::vector<double> e =
     {
         1,5,0,
         -3,2,7,
@@ -132,7 +132,7 @@ TEST_CASE("A submatrix of a 3x3 is a 2x2")
 
 TEST_CASE("A submatrix of a 4x4 is a 3x3")
 {
-    const std::vector<float> e =
+    const std::vector<double> e =
     {
         1,2,3,4,
         5,6,7,8,
@@ -146,7 +146,7 @@ TEST_CASE("A submatrix of a 4x4 is a 3x3")
 
 TEST_CASE("Calculate a minor of a 3x3 matrix")
 {
-    const std::vector<float> e =
+    const std::vector<double> e =
     {
         3,5,0,
         2,-1,-7,
@@ -163,7 +163,7 @@ TEST_CASE("Calculate a minor of a 3x3 matrix")
 
 TEST_CASE("Calculate the cofactor of a 3x3 matrix") {
     // Given the following 3x3 matrix A:
-    const std::vector<float> e = {
+    const std::vector<double> e = {
         3, 5, 0,
         2, -1, -7,
         6, -1, 5
@@ -185,7 +185,7 @@ TEST_CASE("Calculate the cofactor of a 3x3 matrix") {
 
 TEST_CASE("Calculate the cofactor and determinant of a 3x3 matrix") {
     // Given the following 3x3 matrix A:
-    const std::vector<float> e = {
+    const std::vector<double> e = {
         1, 2, 6,
         -5, 8, -4,
         2, 6, 4
@@ -207,7 +207,7 @@ TEST_CASE("Calculate the cofactor and determinant of a 3x3 matrix") {
 
 TEST_CASE("Calculate the cofactor and determinant of a 4x4 matrix") {
     // Given the following 4x4 matrix A:
-    const std::vector<float> e = {
+    const std::vector<double> e = {
         -2, -8,  3,  5,
         -3,  1,  7,  3,
          1,  2, -9,  6,
@@ -233,7 +233,7 @@ TEST_CASE("Calculate the cofactor and determinant of a 4x4 matrix") {
 
 TEST_CASE("Testing an invertible matrix for invertibility") {
     // Given the following 4x4 matrix A:
-    const std::vector<float> e = {
+    const std::vector<double> e = {
         6,  4,  4,  4,
         5,  5,  7,  6,
         4, -9,  3, -7,
@@ -250,7 +250,7 @@ TEST_CASE("Testing an invertible matrix for invertibility") {
 
 TEST_CASE("Testing a noninvertible matrix for invertibility") {
     // Given the following 4x4 matrix A:
-    const std::vector<float> e = {
+    const std::vector<double> e = {
         -4,  2, -2, -3,
          9,  6,  2,  6,
          0, -5,  1, -5,
@@ -267,7 +267,7 @@ TEST_CASE("Testing a noninvertible matrix for invertibility") {
 
 TEST_CASE("Calculating the inverse of a matrix") {
     // Given the following 4x4 matrix A:
-    const std::vector<float> e = {
+    const std::vector<double> e = {
         -5,  2,  6, -8,
          1, -5,  1,  8,
          7,  7, -6, -7,
@@ -294,7 +294,7 @@ TEST_CASE("Calculating the inverse of a matrix") {
     REQUIRE(b.element(2, 3) == Catch::Approx(105.0 / 532.0).epsilon(0.001));
 
     // And B should be the following 4x4 matrix:
-    const std::vector<float> expected = {
+    const std::vector<double> expected = {
         0.21805, 0.45113, 0.24060, -0.04511,
        -0.80827, -1.45677, -0.44361, 0.52068,
        -0.07895, -0.22368, -0.05263, 0.19737,
@@ -307,7 +307,7 @@ TEST_CASE("Calculating the inverse of a matrix") {
 
 TEST_CASE("Calculating the inverse of more matrices") {
     SECTION("First matrix") {
-        const std::vector<float> e1 = {
+        const std::vector<double> e1 = {
             8, -5, 9, 2,
             7, 5, 6, 1,
             -6, 0, 9, 6,
@@ -315,7 +315,7 @@ TEST_CASE("Calculating the inverse of more matrices") {
         };
         const Matrix A(e1);
 
-        const std::vector<float> expected_inverse1 = {
+        const std::vector<double> expected_inverse1 = {
             -0.15385, -0.15385, -0.28205, -0.53846,
             -0.07692, 0.12308, 0.02564, 0.03077,
             0.35897, 0.35897, 0.43590, 0.92308,
@@ -329,7 +329,7 @@ TEST_CASE("Calculating the inverse of more matrices") {
     }
 
     SECTION("Second matrix") {
-        const std::vector<float> e2 = {
+        const std::vector<double> e2 = {
             9, 3, 0, 9,
             -5, -2, -6, -3,
             -4, 9, 6, 4,
@@ -337,7 +337,7 @@ TEST_CASE("Calculating the inverse of more matrices") {
         };
         const Matrix A(e2);
 
-        const std::vector<float> expected_inverse2 = {
+        const std::vector<double> expected_inverse2 = {
             -0.04074, -0.07778, 0.14444, -0.22222,
             -0.07778, 0.03333, 0.36667, -0.33333,
             -0.02901, -0.14630, -0.10926, 0.12963,
@@ -353,7 +353,7 @@ TEST_CASE("Calculating the inverse of more matrices") {
 
 TEST_CASE("Multiplying a product by its inverse") {
     // Define matrix A
-    const std::vector<float> eA = {
+    const std::vector<double> eA = {
         3, -9,  7,  3,
         3, -8,  2, -9,
         -4,  4,  4,  1,
@@ -362,7 +362,7 @@ TEST_CASE("Multiplying a product by its inverse") {
     const Matrix A(eA);
 
     // Define matrix B
-    const std::vector<float> eB = {
+    const std::vector<double> eB = {
         8,  2,  2,  2,
         3, -1,  7,  0,
         7,  0,  5,  4,
